Split in-memory file cache into its own module
